AAP-90164 Move assignment AS helpers into rbac.py; drop remote handling
_record_role_assignment_activity_stream and _prefetch_assignment_content_objects
were left in signals.py but are no longer connected to any Django signal — their
only callers are the two DAB bulk handlers in awx.main.models.rbac. Move both
next to those handlers so the role-assignment logic lives in one place, and drop
the now-redundant cross-module lazy imports. The activity-stream infrastructure
they lean on (activity_stream_enabled, get_activity_stream_class,
get_current_user_or_none, emit_activity_stream_change) stays in signals.py and is
imported lazily to avoid a models<->signals import cycle.

Also drop the remote/federated content-type branch from the prefetch helper:
AWX has no remote objects, so get_remote_base_class and the issubclass guard
were dead weight.

Remove the orphaned functions and the now-unused RoleUserAssignment import from
signals.py.
